requirements.txt
torch>=2.0.0
torchvision
timm
pyyaml
scikit-learn